+++ lib/widgets/searchable_dropdown.dart
@@ -110,7 +110,7 @@ class SearchableDropdown extends StatelessWidget {
                 if (showBottomButton && multiSelect) ...[
                   Btn(
                       text: confirmButtonText,
-                      onPressed: (fn)=> controller.onSavePressed(),
+                      onPressed:  controller.onSavePressed,
                       width: bottomButtonWidth ?? Get.width * 0.8,
                       bgColor: bottomButtonColor,
                       hasBorder: bottomButtonHasBorder ?? false,