+++ lib/widgets/snack_bar.dart
@@ -5,14 +5,12 @@ import 'package:my_widgets/widgets/txt.dart';
 
 /// Flutter code sample for [Overlay].
 
+class ShowSnackBar {
+  static OverlayEntry? overlayEntry;
 
-
-
-class ShowSnackBar  {
-   OverlayEntry? overlayEntry;
-   Future<void> createHighlightOverlay({
-    required String text,
-     BuildContext? context,
+  static Future<void> createHighlightOverlay({
+    String? text,
+    BuildContext? context,
     Color? bgColor,
     Color? borderColor,
     Color? textColor,
@@ -26,8 +24,8 @@ class ShowSnackBar  {
     EdgeInsets? margin,
     Alignment? alignment,
     Function()? onClosedTap,
-   Widget? body,
-
+    bool autoClose = true,
+    Widget? body,
   }) async {
     // Remove the existing OverlayEntry.
     removeHighlightOverlay();
@@ -43,57 +41,63 @@ class ShowSnackBar  {
         // relative to the NavigationBar destination.
         return SafeArea(
           child: Align(
-            alignment: alignment??Alignment.topRight,
+            alignment: alignment ?? Alignment.topRight,
             heightFactor: 1.0,
-            child: body??Container(
-              decoration: pBoxDecoration(
-                color: bgColor??Colors.blueAccent,
-                radius: radius??6,
-                borderColor: borderColor,
-                hasBorder: borderColor !=null,
-              ),
-              padding: padding??const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
-              margin: margin??const EdgeInsets.all(30),
-              constraints:  BoxConstraints(
-                minWidth: Get.width * 0.4,
-                maxHeight: Get.width * 0.8,
-                minHeight: 0.0,
-                maxWidth: Get.height * 0.8
-              ),
-              height: height== null?null: (height+1),
-              child: Row(
-                mainAxisSize: MainAxisSize.min,
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Flexible(child: Txt(text, textColor: textColor,)),
-                    if(hasCloseIcon)
-                      IconButton(
-                        onPressed: onClosedTap??(){
-                        removeHighlightOverlay();
-                        },
-                        icon: const Icon(Icons.close),
-                        visualDensity: VisualDensity.compact,
-                      )
-                  ],
-                ),
-            ),
-
+            child: body ??
+                Container(
+                  decoration: pBoxDecoration(
+                    color: bgColor ?? Colors.blueAccent,
+                    radius: radius ?? 6,
+                    borderColor: borderColor,
+                    hasBorder: borderColor != null,
                   ),
+                  padding: padding ??
+                      const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+                  margin: margin ?? const EdgeInsets.all(30),
+                  constraints: BoxConstraints(
+                      minWidth: Get.width * 0.4,
+                      maxHeight: Get.width * 0.8,
+                      minHeight: 0.0,
+                      maxWidth: Get.height * 0.8),
+                  height: height == null ? null : (height + 1),
+                  child: Row(
+                    mainAxisSize: MainAxisSize.min,
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Flexible(
+                          child: Txt(
+                        text,
+                        textColor: textColor,
+                      )),
+                      if (hasCloseIcon)
+                        IconButton(
+                          onPressed: onClosedTap ??
+                              () {
+                                removeHighlightOverlay();
+                              },
+                          icon: const Icon(Icons.close),
+                          visualDensity: VisualDensity.compact,
+                        )
+                    ],
+                  ),
+                ),
+          ),
         );
       },
     );
 
     // Add the OverlayEntry to the Overlay.
-    Overlay.of(context??Get.context!).insert(overlayEntry!);
-    await Future.delayed(duration);
-    removeHighlightOverlay();
+    Overlay.of(context ?? Get.context!).insert(overlayEntry!);
+    if (autoClose) {
+      await Future.delayed(duration);
+      removeHighlightOverlay();
+    }
   }
 
   // Remove the OverlayEntry.
-   void removeHighlightOverlay() {
+  static void removeHighlightOverlay() {
     overlayEntry?.remove();
     overlayEntry?.dispose();
     overlayEntry = null;
   }
-
 }